Update celebrate her page (#288)
* Update celebrate her page
---------
