+++ src/main/java/com/wcc/platform/domain/cms/pages/aboutus/CelebrateHerPage.java
@@ -1,5 +1,6 @@
 package com.wcc.platform.domain.cms.pages.aboutus;
 
+import com.wcc.platform.domain.cms.attributes.CommonSection;
 import com.wcc.platform.domain.cms.attributes.HeroSection;
 import com.wcc.platform.domain.platform.AboutHer;
 import jakarta.validation.constraints.NotNull;
@@ -7,4 +8,7 @@ import java.util.List;
 
 /** CMS Celebrate Her page. */
 public record CelebrateHerPage(
-    @NotNull String id, @NotNull HeroSection heroSection, @NotNull List<AboutHer> items) {}
+    @NotNull String id,
+    @NotNull HeroSection heroSection,
+    @NotNull CommonSection section,
+    @NotNull List<AboutHer> items) {}
