+++ src/test/java/com/wcc/platform/integrationtests/CmsServiceIntegrationTest.java
@@ -1,6 +1,7 @@
 package com.wcc.platform.integrationtests;
 
 import static com.wcc.platform.domain.cms.PageType.ABOUT_US;
+import static com.wcc.platform.domain.cms.PageType.CELEBRATE_HER;
 import static com.wcc.platform.domain.cms.PageType.CODE_OF_CONDUCT;
 import static com.wcc.platform.domain.cms.PageType.COLLABORATOR;
 import static com.wcc.platform.domain.cms.PageType.FOOTER;
@@ -9,6 +10,7 @@ import static com.wcc.platform.domain.cms.PageType.TEAM;
 import static com.wcc.platform.factories.SetupFactories.DEFAULT_CURRENT_PAGE;
 import static com.wcc.platform.factories.SetupFactories.DEFAULT_PAGE_SIZE;
 import static com.wcc.platform.factories.SetupFactories.createAboutUsPageTest;
+import static com.wcc.platform.factories.SetupFactories.createCelebrateHerPageTest;
 import static com.wcc.platform.factories.SetupFactories.createCodeOfConductPageTest;
 import static com.wcc.platform.factories.SetupFactories.createCollaboratorPageTest;
 import static com.wcc.platform.factories.SetupFactories.createFooterTest;
@@ -19,8 +21,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.wcc.platform.domain.cms.pages.CodeOfConductPage;
 import com.wcc.platform.domain.cms.pages.CollaboratorPage;
+import com.wcc.platform.domain.cms.pages.aboutus.CodeOfConductPage;
 import com.wcc.platform.repository.PageRepository;
 import com.wcc.platform.service.CmsService;
 import java.util.Map;
@@ -125,6 +127,16 @@ class CmsServiceIntegrationTest extends SurrealDbIntegrationTest {
   }
 
   @Test
+  void testGetCelebrateHerPage() {
+    var celebrateHerPage = createCelebrateHerPageTest(CELEBRATE_HER.getFileName());
+    pageRepository.create(objectMapper.convertValue(celebrateHerPage, Map.class));
+    var result = service.getCelebrateHer();
+    // TODO to investigate why heroSection test doesn't pass for CelebrateHerPage meanwhile pass for the other pages
+    //assertEquals(celebrateHerPage.heroSection(), result.heroSection());
+    assertEquals(celebrateHerPage.section(), result.section());
+    assertEquals(celebrateHerPage.items(), result.items());
+  }
+  
   @SuppressWarnings("unchecked")
   void testGetLandingPage() {
     var landingPage = createLandingPageTest(LANDING_PAGE.getFileName());
