--- src/test/resources/celebrateHerPage.json
@@ -1,29 +1,0 @@
-{
-  "id": "page:CELEBRATE_HER",
-  "heroSection": {
-    "title": "Celebrate Her",
-    "images": [
-      {
-        "path": "https://cloudprovider.com/image.png",
-        "alt": "There is a group of women showing WCC logo",
-        "type": "desktop"
-      }
-    ]
-  },
-  "section": {
-    "description": "Women Coding Community highlights the achievements of our members by celebrating them with Celebrate Her"
-  },
-  "items": [
-    {
-      "listOfName": [
-        "Liliia Rafikova"
-      ],
-      "description": "Celebrate her for first commit to wcc backend.",
-      "link": {
-        "title": "",
-        "type": "linkedIn",
-        "link": "https://www.linkedin.com/company/womencodingcommunity"
-      }
-    }
-  ]
-}
